+++ kasane-plugin-package/src/package.rs
@@ -463,7 +463,7 @@ mod tests {
             r#"
 [plugin]
 id = "demo_plugin"
-abi_version = "4.2.0"
+abi_version = "5.0.0"
 
 [capabilities]
 wasi = ["filesystem", "process"]